hyperstructure level: update combat burn and steal

+++ contracts/src/models/hyperstructure.cairo
@@ -5,7 +5,7 @@ struct HyperStructure {
     hyperstructure_type: u8,
     construction_resource_id: u128,
     construction_resource_count: u32,
-    order: u32,
+    order: u8,
     level: u32,
     max_level: u32
 }
